+++ valiktor-spring/valiktor-spring-boot-autoconfigure/src/main/kotlin/org/valiktor/springframework/boot/autoconfigure/ValiktorWebFluxAutoConfiguration.kt
@@ -16,6 +16,7 @@
 
 package org.valiktor.springframework.boot.autoconfigure
 
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.context.annotation.Bean
@@ -32,10 +33,11 @@ import org.valiktor.springframework.web.reactive.ValiktorReactiveExceptionHandle
  * @since 0.3.0
  */
 @Configuration
-@ConditionalOnClass(name = [
+@ConditionalOnClass(name = ["org.springframework.web.reactive.DispatcherHandler"])
+@ConditionalOnBean(type = [
     "org.valiktor.springframework.config.ValiktorConfiguration",
-    "org.springframework.web.reactive.DispatcherHandler",
-    "org.springframework.http.codec.CodecConfigurer"])
+    "org.springframework.http.codec.CodecConfigurer"
+])
 class ValiktorWebFluxAutoConfiguration {
 
     /**